Added documentation folder
We need to document our work so I added the documentation folder. I added a file with documentation on the frequencies of the notes on a guitar finger board. Also added images of the simulink diagram and the values used for the filtering. I took 2 screen shots of the Kaiser and Chebishev filtering.